Adds cvars sys_homedir and sys_libdir to get q2pro default directories.
This cvars will be used in GetGameAPI() of g_main.c to try to load the
game DLL from the home or library directory. If all fails, it will try
to load from the current directory (previous behavoir).

Also, adds macro defines to avoid double inclusion of header files and
adds __$ARCH__ define to CFLAGS in GNUmakefile.

Q2ADMINVERSION is now 1.17.48-kmod-3